Refactor input (#56)
* removes detection and conversion packages

* split cli package into multiple files

---------
