+++ internal/cli/cli.go
@@ -4,14 +4,6 @@ import (
 	"flag"
 	"log"
 	"os"
-
-	"github.com/ChainsAre2Tight/qr-decoder-utils/internal/conversion"
-	"github.com/ChainsAre2Tight/qr-decoder-utils/internal/decoding"
-	qr "github.com/ChainsAre2Tight/qr-decoder-utils/internal/decoding/QR"
-	"github.com/ChainsAre2Tight/qr-decoder-utils/internal/detection"
-	"github.com/ChainsAre2Tight/qr-decoder-utils/internal/input"
-	"github.com/ChainsAre2Tight/qr-decoder-utils/internal/output"
-	"github.com/ChainsAre2Tight/qr-decoder-utils/internal/utils"
 )
 
 var (
@@ -56,100 +48,4 @@ func CLI() {
 		printUsage()
 	}
 
-}
-
-func requireInputName() {
-	if *inputFilenamePtr == "" {
-		log.Print("Input filename not specified")
-		printUsage()
-	}
-}
-
-// will generate random filename if flag wasn't provided
-func requireOutputName() {
-	if *outputFilenamePtr == "" {
-		*outputFilenamePtr = utils.GenerateRandomFilename()
-		log.Printf("WARNING: No --output filename specified, writting to %s", *outputFilenamePtr)
-	}
-}
-func validateOutputSize() {
-	if *outputSizePtr < 0 || *outputSizePtr > 100 {
-		log.Print("Output size out of range.")
-		printUsage()
-	}
-}
-
-func decode() {
-	requireInputName()
-	matrix := LoadAndConvert(inputFilenamePtr)
-
-	data, err := decoding.Decode(matrix)
-	if err != nil {
-		log.Fatal(err)
-	}
-	log.Println("Decoded contents:\n\n", data)
-}
-
-func convertImage() {
-	requireInputName()
-	requireOutputName()
-	matrix := LoadAndConvert(inputFilenamePtr)
-	output.MatrixToImage(matrix, *outputFilenamePtr)
-}
-
-func LoadAndConvert(inputFilenamePtr *string) [][]bool {
-
-	// load image
-	img := input.ReadImage(*inputFilenamePtr)
-
-	// detect borders and resize
-	validateOutputSize()
-	qr, err := detection.DetectCodeOnImage(img, *outputSizePtr)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	// convert to matrix
-	matrix := conversion.ImageToMartix(qr)
-	return matrix
-}
-
-func convertExcel() {
-	requireInputName()
-	requireOutputName()
-
-	var outputFunction func([][]bool, string)
-	if *includeMasksPtr {
-		outputFunction = output.MatrixToExcelWithMasks
-	} else {
-		outputFunction = output.MatrixToExcel
-	}
-
-	matrix := LoadAndConvert(inputFilenamePtr)
-
-	// output to selected format
-	outputFunction(matrix, *outputFilenamePtr)
-}
-
-func mask() {
-	requireOutputName()
-	validateOutputSize()
-
-	mask, ok := qr.Masks[*maskPtr]
-	if !ok {
-		log.Printf("Mask \"%s\" is unknown", *maskPtr)
-		printUsage()
-	}
-
-	if *outputSizePtr == 0 {
-		*outputSizePtr = 21
-	}
-	result := qr.GenerateMaskedMatrix(*outputSizePtr, mask)
-
-	output.MatrixToExcel(result, *outputFilenamePtr)
-}
-
-func printUsage() {
-	flag.PrintDefaults()
-	os.Exit(1)
 }
